backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
openai>=1.3.7
python-dotenv==1.0.0
pydantic>=2.5.0
python-multipart==0.0.6
fastapi-cors==0.0.6 